feat: add PolicyCompat component
