+++ src/components/PolicyCompat.astro
@@ -1,0 +1,117 @@
+---
+import schema from "../../schema/policies-schema.json";
+
+interface Props {
+  /** Match a key under `properties` in policies-schema.json. */
+  policy: string;
+}
+
+const { policy } = Astro.props;
+
+// Fixed display order, so the table doesn't depend on key order in the schema.
+const CHANNELS = [
+  { key: "firefox", label: "Firefox" },
+  { key: "firefox_esr", label: "Firefox ESR" },
+  { key: "firefox_enterprise", label: "Firefox Enterprise" },
+] as const;
+
+// The version replaces `visible` if the channel is supported.
+const STATES = {
+  yes: { announced: "Available since ", visible: null },
+  no: { announced: "Not supported", visible: "No" },
+} as const;
+
+type State = keyof typeof STATES;
+type Support = { version_added?: string | false };
+type Entry = { "x-compatibility"?: Record<string, Support | undefined> };
+
+const compatibility = (schema.properties as Record<string, Entry>)[policy]?.["x-compatibility"];
+
+const cells = CHANNELS.map(({ key, label }) => {
+  const added = compatibility?.[key]?.version_added;
+  const state: State = typeof added === "string" ? "yes" : "no";
+  return { label, state, version: typeof added === "string" ? added : null };
+});
+
+// A policy missing from the schema, or one with no `x-compatibility`
+if (!compatibility) {
+  console.warn(`[PolicyCompat] no x-compatibility for "${policy}" in policies-schema.json`);
+}
+---
+
+{
+  compatibility && (
+    <table class="compat">
+      <thead>
+        <tr>
+          {cells.map(({ label }) => (
+            <th scope="col">{label}</th>
+          ))}
+        </tr>
+      </thead>
+      <tbody>
+        <tr>
+          {cells.map(({ state, version }) => (
+            <td class={state}>
+              <span class="cell">
+                <svg
+                  viewBox="0 0 16 16"
+                  width="14"
+                  height="14"
+                  aria-hidden="true"
+                  fill="none"
+                  stroke="currentColor"
+                  stroke-width={state === "yes" ? 2 : 1.5}
+                  stroke-linecap="round"
+                  stroke-linejoin="round"
+                >
+                  {state === "yes" ? (
+                    <path d="M2.5 8.5 6 12l7.5-8" />
+                  ) : (
+                    <>
+                      <circle cx="8" cy="8" r="6.25" />
+                      <path d="m5.75 5.75 4.5 4.5m0-4.5-4.5 4.5" />
+                    </>
+                  )}
+                </svg>
+                <span class="sr-only">{STATES[state].announced}</span>
+                {version ?? <span aria-hidden="true">{STATES[state].visible}</span>}
+              </span>
+            </td>
+          ))}
+        </tr>
+      </tbody>
+    </table>
+  )
+}
+
+<style>
+  .compat {
+    width: fit-content;
+    border: 1px solid #51565d;
+    border-radius: 0.25rem;
+  }
+  .compat :is(th, td) {
+    white-space: nowrap;
+  }
+
+  table.compat :is(th, td):is(:first-child, :last-child) {
+    padding-inline: 1rem;
+  }
+  /* The frame draws the bottom edge, so drop the last row's cell rule. */
+  .compat tbody tr:last-child :is(th, td) {
+    border-bottom: 0;
+  }
+  .cell {
+    display: inline-flex;
+    align-items: center;
+    gap: 0.375rem;
+  }
+  /* The `-high` tokens clear WCAG AA against the page background(s) */
+  .yes {
+    color: var(--sl-color-green-high);
+  }
+  .no {
+    color: var(--sl-color-red-high);
+  }
+</style>
