+++ src/content/docs/reference/policies/DefaultBrowserSettingEnabled.mdx
@@ -11,7 +11,10 @@ When it is set to `true`, the default browser check runs at startup and users ca
 
 To suppress the startup check without removing the controls from browser preferences, use the [`DontCheckDefaultBrowser`](/reference/policies/dontcheckdefaultbrowser/) policy instead.
 
-**Compatibility:** Firefox 154\
+## Compatibility
+
+<PolicyCompat policy="DefaultBrowserSettingEnabled" />
+
 **CCK2 Equivalent:** N/A\
 **Preferences Affected:** `browser.shell.checkDefaultBrowser`
